V2/requirements.txt
pymupdf
python-docx
fastapi
uvicorn
sentence-transformers
faiss-cpu
